[LifetimeSafety] Refactor assignment history lookup into AssignmentQuery
Independent AssignmentQuery now handles the logic for searching  assignment history during error reporting, decoupling it from the diagnostic logic.
